requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.30.6
git+https://github.com/coinbase/x402.git#subdirectory=python/x402
python-dotenv>=1.0.0
jinja2>=3.1.4
